requirements.txt
asgiref==3.8.1
branca==0.7.2
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-rest-framework==0.1.0
djangorestframework==3.15.2
fastkml==0.12
folium==0.17.0
geojson==3.1.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
pygeoif==0.7
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==70.3.0
shapely==2.0.5
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
xyzservices==2024.6.0
